[rebroadcast] Track rebroadcast attempts
Co-authored-by: Will Clark <will8clark@gmail.com>

Introduce a boost::multi_index to track our previous rebroadcast attempts. We
use this information primarily to ensure we don't exceed a maximum number of
rebroadcast attempts. This allows transactions to genuinely expire from the
majority of mempools. We disable automatic rebroadcasts because the
conceptually desired behavior should be that after a certain amount of time,
only the source wallet should rebroadcast the transaction. (Of course, we
cannot enforce this on a decentralized network, but design the default behavior
to uphold this idea.)