[rebroadcast] Stop tracking rebroadcast attempts for certain transactions
Co-authored-by: Will Clark <will8clark@gmail.com>

Depending on the reason a transaction is removed from the mempool, it can be
extremely unlikely for it to be able to re-enter. Some examples are if it was
mined into a block or replaced using RBF. In these circumstances, we can remove
the transaction entry from the rebroadcast tracker. Under an unlikely
circumstance where the transaction does re-enter the mempool and gets
rebroadcast, it will simply be re-added to the attempt tracker.