amend unspecified pointer operator comparison
fixes #19496

> If two pointers are not specified to compare greater or compare equal,
the result of the comparison is unspecified. The result may be
nondeterministic, and need not be consistent even for multiple
evaluations of the same expression with the same operands in the same
execution of the program.

> Pointers are only specified to have a legal < or > comparison
operations if the objects they point to either belong to the same array,
or if the objects they point to live basically as members in the same
object (i.e. as class members).

This behaviour has not been observed occurring in practice in Bitcoin
Core to-date.