net: Prevent discard final msg after disconnect
After a clean remote disconnection, process one final message if it
exists.

Previously it was possible for a message to be received followed by a
clean TCP disconnect. At this point there exists a race between
ProcessMessages() on ThreadMessageHandler and DisconnectNodes() on
ThreadSocketHandler to determine whether the final message would end up being
processed or not.

This commit adds a new bool fDisconnectNext which pre-marks the node for
disconnection. ThreadMessageHandler will then handle one more message if
it exists, before fully marking the node for disconnection by setting
fDisconnect, so the node can be removed by DisconnectNodes as normal.