lint: add markdown link checker
This adds a broken markdown link check job to the lint test_runner.
It relies on having the [`mlc`](https://crates.io/crates/mlc/0.16.3)
binary found on $PATH, but will fail with `success` if the binary is not
found.

`mlc` is also added to the ci/04_install.sh script run by the
containerfile.