validation: Use span for CalculateClaimedHeadersWork
Avoids the need for copying by having to wrap the header in a
std::vector when just a single header is passed. Also makes it
friendlier for potential future users of the kernel library if they do
not store the headers in a std::vector, but can guarantee contiguous
memory.