ci: add native_valgrind as nightly CI job
Reintroduce the native_valgrind CI job on a nightly schedule.

As this job takes longer than the others, and is not currently being run
at all, don't nerf CI runtime by running it on every push/pull_request.

Instead run it as a nightly job, or on workflow dispatch (manually).